daemon: start moving implementation to api_snaps.go
and api_sideload_n_try.go

for now this moves many snaps including querying about them

and sideload/try code, in the process create a separate
sideloadOrTrySnap

and simplify the signature of trySnap

two renames as well:

snapsOp => snapManyOp
snapsActionFunc => snapManyActionFunc
